fix: ensure repo naddr consistancy
proposals / issues are fetched that tag the repo address pointer
for each maintainer listed in the selected repo event. not just the
ones that tag the selected proposal event.

previously, the link to each proposal would change the selected
repository to the first address pointer tagged in the proposal event

not only is the selected repo changing without the user's knowledge,
it can also cause an error if it points to a repo event that doesn't
exist. this happned here:
nostr:note1h25kwpycszl33285exdkrm4qz7a03e37nf4ujvexwqz94yj3k4xql9yj23

this ensures that the selected repo address pointer doesnt change.

if the proposal is accessed via gitworkhop.dev/e/<ref> the page will
error. a seperate change is needed to check through each
address pointer listed until it finds one that points to an actual
event.