fix: show thread on pr and issue page
from the in memory database which should be populated by the
local cache db

this might get pruned in the future so we should fetch it again.

the plan was to store the ids of responses in the issue / pr so
maybe we should check whether those ids are present in the cache
and if not fetch from relays.