enable open subscriptions
enable query_centre queries to be called multiple times but use a
single open suscription that auto cloes when all subscribers
components are distroyed. its amazing how much easier this would be
with svelte 4 stores.

main use case is fetchRepo which is included in this commit but can
apply to fetching current user data and other pages such as
user page.