feat(actions): listen for job offers
add listenForActions.

this is not tested but I wanted to commit this on the off-chance
Arjen wanted to take a look as he is in another timezone.

we are reusing the watchRepo subscription to avoid duplication and
because it doesn't really matter if we ask for a few extra kinds.

fetchActions needs more thought. do we want actions related to
current state and recently run actions? we should manage the 'since'
a bit better and do it as part of normal fetchRepo.