fix: user header flashing when parent updated
this approach fixes the scenario where the component is remounted
repeatedly with the same prop value.

its often because the prop comes from an object that has other
properties that are being updated.

there is a async function that gets the user details and this caused
the flash in this instance.