feat: show issue or pr mentions
before deletion was added mentions appeared as replies in the thread.
this was a bit confusing as they read differently if included in a
thread than seen as just mentioning it.

when deletions were added and threading reworked to show where
a reply parent was missing, mentions completely disappeared

now they are added back in but just without actually showing the
content. where they are mentioned in a reply (often may be a issue
or a PR) the root event is also linked.

the query_centre fetchEvent function needs improvement and so does
routing of links to specific events that aren't Prs, issues or
repo anns.