fix: blank PR thread
`thread_tree.ts` function `getRootId` was overloaded and used for too
many different purposes.

`getRootUuid` will never return the id of the event passed to it and
will return undefined in the event doesn't list a root.

the complexities of finding the the root in an absolute sense is
given to the nip10 and nip22 thread tag generation functions