feat(explorer): commits page WIP
Note that routes `<repo>/commits` and `<repo>/prs/<id>/commits`
overlap so that the pr one doesnt resolve. look into this. for now
we just omitted the 's'.

I think maybe it would be better to undo the git-manager changes and
replace with a eventmesage that gets trigged when selectedRef is
updated that emmits string[] of commit ids younges to oldest.

If we are worried about space for larger repos we could use the
shorthand and evauate their length. and expand them using isomophic
git if we need to load the commit info.